Avances en crud mesas v05-10-22

+++ restaurant-siglo-xxi/src/main/java/com/restaurant/siglo/xxi/repository/EstadoMesaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.restaurant.siglo.xxi.clases.EstadoMesa;
 
-public interface EstadoMesaRepository extends JpaRepository<EstadoMesa, Long>{
+public interface EstadoMesaRepository extends JpaRepository<EstadoMesa, Integer>{
